refactor: 점수 계산 책임 Board로 변경
