+++ src/main/java/domain/game/Board.java
@@ -1,14 +1,17 @@
 package domain.game;
 
-import domain.piece.EmptyPiece;
-import domain.piece.Piece;
-import domain.piece.Position;
-import domain.piece.Side;
+import domain.piece.*;
 
+import java.util.Arrays;
+import java.util.EnumMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 public class Board {
+
+    private static final double DUPLICATE_PAWN_SCORE = 0.5d;
+
     private final Map<Position, Piece> chessBoard;
 
     public Board(Map<Position, Piece> chessBoard) {
@@ -31,6 +34,18 @@ public class Board {
 
     public boolean isKing(Position targetPosition) {
         return chessBoard.get(targetPosition).isSameType(PieceType.KING);
+    }
+
+    public Map<Side, Double> calculateScore() {
+        Map<Side, Double> scores = new EnumMap<>(Side.class);
+
+        List<Piece> whitePieces = collectPiecesBySide(Side.WHITE);
+        List<Piece> blackPieces = collectPiecesBySide(Side.BLACK);
+
+        scores.put(Side.WHITE, calculateScoreOneSidePieces(whitePieces) - calculateMinusScoreBySameRankPawn(Side.WHITE));
+        scores.put(Side.BLACK, calculateScoreOneSidePieces(blackPieces) - calculateMinusScoreBySameRankPawn(Side.BLACK));
+
+        return scores;
     }
 
     private void validateTurn(Side side, Piece sourcePiece) {
@@ -69,6 +84,30 @@ public class Board {
     private void movePiece(Position sourcePosition, Position targetPosition, Piece sourcePiece) {
         this.chessBoard.put(sourcePosition, new EmptyPiece());
         this.chessBoard.put(targetPosition, sourcePiece);
+    }
+
+    private List<Piece> collectPiecesBySide(Side side) {
+        return chessBoard.values().stream()
+                .filter(piece -> piece.isSameSide(side))
+                .collect(Collectors.toList());
+    }
+
+    private double calculateScoreOneSidePieces(List<Piece> pieces) {
+        return pieces.stream()
+                .mapToDouble(Piece::score)
+                .sum();
+    }
+
+    private double calculateMinusScoreBySameRankPawn(Side side) {
+        Map<File, Long> duplicatePawns = Arrays.stream(File.values())
+                .flatMap(file -> Arrays.stream(Rank.values()).map(rank -> Position.of(file.getText(), rank.getText())))
+                .filter(position -> chessBoard.get(position).isSameSide(side) && chessBoard.get(position).isSameType(PieceType.PAWN))
+                .collect(Collectors.groupingBy(Position::getFile, Collectors.counting()));
+
+        return duplicatePawns.values().stream()
+                .filter(duplicatePawnCount -> duplicatePawnCount > 1)
+                .mapToDouble(duplicatePawnCount -> duplicatePawnCount * DUPLICATE_PAWN_SCORE)
+                .sum();
     }
 
     public Map<Position, Piece> getChessBoard() {
